+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add manual e2e testing procedure and script.
 
+### Removed
+
+- Remove the alloy-secret resource which is no longer needed as Alloy secret was moved into logging-secret
+
 ## [0.12.1] - 2024-09-23
 
 ### Fixed